--- src/components/MoviesCardList/MoviesCard/MoviesCard.css
@@ -1,105 +1,0 @@
-.movies-card {
-    display: flex;
-    flex-direction: column;
-}
-
-.movies-card__item {
-    overflow: hidden;
-    display: flex;
-    position: relative;
-}
-
-.movies-card__item:last-child {
-    margin: 14px 14px 0;
-    align-items: center;
-    text-align: center;
-    justify-content: space-between;
-}
-
-.movies-card__item_img {
-    border-radius: 6px;
-    width: 100%;
-    object-fit: cover;
-    object-position: center;
-    -moz-transition: all 400ms ease-out;
-    -o-transition: all 400ms ease-out;
-    -webkit-transition: all 400ms ease-out;
-}
-
-.movies-card__item_img:hover {
-    -webkit-transform: scale(1.1);
-    -moz-transform: scale(1.1);
-    -o-transform: scale(1.1);
-}
-
-.movies-card__item_head {
-    margin: 0;
-    font-weight: 500;
-    font-size: 13px;
-    line-height: 16px;
-    text-align: start;
-    max-width: 250px;
-}
-
-.movies-card__item_dur {
-    margin: 0;
-    padding: 3px 4px;
-    font-weight: normal;
-    font-size: 11px;
-    line-height: 10px;
-    color: #8B8B8B;
-    border-radius: 3px;
-    background: rgba(47, 47, 47, 1);
-    white-space: nowrap;
-}
-
-.movies-card__item_save {
-    opacity: 0;
-    top: 14px;
-    right: 14px;
-    position: absolute;
-    border: none;
-    background: #313131;
-    color: white;
-    border-radius: 30px;
-    cursor: pointer;
-}
-
-.movies-card__item:hover .movies-card__item_save {
-    opacity: 1;
-    transition: linear .3s;
-}
-
-.movies-card__item_save:hover {
-    background: #EE3465;
-}
-
-.movies-card__item_saved {
-    width: 21px;
-    height: 21px;
-    top: 14px;
-    right: 14px;
-    position: absolute;
-    border: none;
-    border-radius: 30px;
-    background: #EE3465 url("../../../images/tick_icon.svg") no-repeat center;
-    cursor: pointer;
-}
-
-.movies-card__item_delete {
-    opacity: 0;
-    width: 21px;
-    height: 21px;
-    top: 14px;
-    right: 14px;
-    position: absolute;
-    border: none;
-    border-radius: 30px;
-    background: #313131 url("../../../images/delete_icon.svg") no-repeat center;
-    cursor: pointer;
-}
-
-.movies-card__item:hover .movies-card__item_delete {
-    opacity: 1;
-    transition: linear .3s;
-}
